[12.0][IMP] move all state related code from fatturapa_pec to fatturapa_out
