+++ l10n_it_fatturapa_out/readme/CONTRIBUTORS.rst
@@ -4,6 +4,7 @@
 * Alessio Gerace
 * Alex Comba
 * Sergio Zanchetta <https://github.com/primes2h>
+* Giovanni Serra <giovanni@gslab.it>
 
 * `Tecnativa <https://www.tecnativa.com>`_:
