+++ l10n_it_fatturapa_out/views/attachment_view.xml
@@ -10,6 +10,9 @@
                     <field name="name"/>
                     <field name="invoice_partner_id"/>
                     <field name="out_invoice_ids"/>
+                    <field name="state"/>
+                    <field name="sending_date"/>
+                    <field name="delivered_date"/>
                 </tree>
             </field>
         </record>
@@ -19,6 +22,11 @@
             <field name="arch" type="xml">
                 <search>
                     <field name="name"/>
+                    <field name="state"/>
+                    <filter name="ready" string="Ready to Send" domain="[('state','=','ready')]"/>
+                    <filter name="sender_error" string="Sender Error" domain="[('state','=','sender_error')]"/>
+                    <filter name="recipient_error" string="Not delivered" domain="[('state','=','recipient_error')]"/>
+                    <filter name="rejected" string="Rejected (PA)" domain="[('state','=','rejected')]"/>
                     <field name="invoice_partner_id"/>
                 </search>
             </field>
@@ -37,7 +45,8 @@
                             <group>
                                 <label for="datas"/>
                                 <div>
-                                    <field name="datas" filename="datas_fname"/>
+                                    <field name="datas" filename="datas_fname"
+                                           attrs="{'readonly': [('state', 'not in', ['ready', 'sender_error'])]}"/>
                                     <field name="ftpa_preview_link" widget="url" text="Show preview"/>
                                 </div>
                                 <field name="datas_fname" />
@@ -86,6 +95,7 @@
             <field name="view_type">form</field>
             <field name="view_mode">tree,form</field>
             <field name="view_id" ref="view_fatturapa_out_attachment_tree"/>
+            <field name="context">{'search_default_ready': True, 'search_default_sender_error': True, 'search_default_rejected': True, }</field>
         </record>
 
         <menuitem action="action_fatturapa_attachment"
