+++ l10n_it_fatturapa_out/wizard/__init__.py
@@ -8,5 +8,6 @@ _logger = logging.getLogger(__name__)
 # in wizard_export_fatturapa, so we guard the entire file
 try:
     from . import wizard_export_fatturapa
+    from . import wizard_export_fatturapa_regenerate
 except ImportError:
     _logger.debug('Cannot `import pyxb`.')  # Avoid init error if not installed
