+++ l10n_it_fatturapa_out/wizard/wizard_export_fatturapa_regenerate.py
