+++ l10n_it_fatturapa_pec/models/__init__.py
@@ -1,5 +1,4 @@
 
-from . import account
 from . import fatturapa_attachment_out
 from . import mail_thread
 from . import fetchmail
