+++ l10n_it_fatturapa_pec/models/fatturapa_attachment_out.py
@@ -24,8 +24,6 @@ class FatturaPAAttachmentOut(models.Model):
     last_sdi_response = fields.Text(
         string='Last Response from Exchange System', default='No response yet',
         readonly=True)
-    sending_date = fields.Datetime("Sent Date", readonly=True)
-    delivered_date = fields.Datetime("Delivered Date", readonly=True)
     sending_user = fields.Many2one("res.users", "Sending User", readonly=True)
 
     @api.model
