+++ l10n_it_fatturapa_pec/readme/CONTRIBUTORS.rst
@@ -2,6 +2,7 @@
 * Sergio Corato <info@efatto.it>
 * Lorenzo Battistini <https://github.com/eLBati>
 * Sergio Zanchetta <https://github.com/primes2h>
+* Giovanni Serra <giovanni@gslab.it>
 
 * `Tecnativa <https://www.tecnativa.com>`_:
