+++ l10n_it_fatturapa_pec/views/fatturapa_attachment_out.xml
@@ -15,9 +15,6 @@
                     <field name="last_sdi_response" nolabel="1"/>
                 </group>
             </xpath>
-            <field name="datas" position="attributes">
-                <attribute name="attrs">{'readonly': [('state', 'not in', ['ready', 'sender_error'])]}</attribute>
-            </field>
             <div name="creation_div" position="after">
                 <label for="sending_user" string="Sent by" attrs="{'invisible': [('sending_user' ,'=', False)]}"/>
                 <div name="sending_user" attrs="{'invisible': [('sending_user' ,'=', False)]}">
@@ -44,35 +41,5 @@
         </field>
     </record>
 
-    <record model="ir.ui.view" id="view_fatturapa_out_attachment_tree_state">
-        <field name="name">view_fatturapa_out_attachment_tree_state</field>
-        <field name="model">fatturapa.attachment.out</field>
-        <field name="inherit_id" ref="l10n_it_fatturapa_out.view_fatturapa_out_attachment_tree"/>
-        <field name="arch" type="xml">
-            <field name="out_invoice_ids" position="after">
-                <field name="state"/>
-                <field name="sending_date"/>
-                <field name="delivered_date"/>
-            </field>
-        </field>
-    </record>
-    <record model="ir.ui.view" id="view_fatturapa_out_attachment_search_state">
-        <field name="name">view_fatturapa_out_attachment_search_state</field>
-        <field name="model">fatturapa.attachment.out</field>
-        <field name="inherit_id" ref="l10n_it_fatturapa_out.view_fatturapa_out_attachment_search"/>
-        <field name="arch" type="xml">
-            <field name="name" position="after">
-                <field name="state"/>
-                <filter name="ready" string="Ready to Send" domain="[('state','=','ready')]"/>
-                <filter name="sender_error" string="Sender Error" domain="[('state','=','sender_error')]"/>
-                <filter name="recipient_error" string="Not delivered" domain="[('state','=','recipient_error')]"/>
-                <filter name="rejected" string="Rejected (PA)" domain="[('state','=','rejected')]"/>
-            </field>
-        </field>
-    </record>
-
-    <record id="l10n_it_fatturapa_out.action_fatturapa_attachment" model="ir.actions.act_window">
-        <field name="context">{'search_default_ready': True, 'search_default_sender_error': True, 'search_default_rejected': True, }</field>
-    </record>
 
 </odoo>
