+++ l10n_it_fatturapa_pec/wizard/__init__.py
@@ -1,3 +1,2 @@
 
 from . import send_pec
-from . import wizard_export_fatturapa
